Documentation updates for the background.js. Added JSDoc strings for functions. Labeled Global listener per documentation suggestions.
The only code that's been reworked that is being utilized is the getRandomSubreddit moving from chained if statements to a switch statement for clarity. I haven't tested it yet because I don't have a test environment setup yet.

I did add a openSubredditInCurrentTab function. The idea is that I can also add a logical statement at some point that allows you to check the users current tab URL => If its on reddit we can imply user consent to navigate from the page => Else open a new tab => Navigate to random page. This would make it so a user can traverse their stumblr session in one tab. Future idea that I don't know is worth it yet.